Extract additional methods
